Make Proposed Solution less redundant and pull in more aspects of the proposal
The formulation of Proposed Solution as a restatement of the motivation
felt redundant, and missed a lot of key aspects of the proposed
solution. Bring some more fundamental ideas of how typed throws works,
such as how `do...catch` works and the relationship with `any Error`
and `Never`-throwing types, into the proposed solution.