Expand the detailed discussion of throwing and catching errors
The detailed discussion for throwing and catching typed errors was a
little sparse. Expand it to explain the rules in more detail, with
additional examples, and give it a structure that separately copes
with "what you can throw" and "how you catch".

Note that there is a significant change to the proposed semantics
here! When the `do` statement contains throw sites that all throw the
same error type, the thrown error type presented to the `catch` blocks
is of that thrown error type. This differs from the proposal, which
kept the thrown error type as `any Error` all the time. The (new)
proposed semantics are a better match for typed throws because it lets
you take advantage of knowledge of the thrown type, and you don't need
to rely on the syntax of dynamic checks like "let e as CatError" in
such cases.