Improve AsyncSequence/AsyncIteratorProtocol and async for...in loops
Typed throws gives us the mechanism to replace the never-formalized
`rethrows` protocols, which were also in the way of adopting primary
associated types in these protocols. Use the `Failure` type of an
`AsyncSequence` to inform asynchronous for loops as to whether (and
what) they can throw, and adopt primary associated types.